chore: deduper refactor
