+++ lib/arborist/build-ideal-tree.js
@@ -33,7 +33,7 @@ const relpath = require('../relpath.js')
 // them with unit tests and reuse them across mixins
 const _complete = Symbol('complete')
 const _depsSeen = Symbol('depsSeen')
-const _depsQueue = Symbol.for('depsQueue')
+const _depsQueue = Symbol('depsQueue')
 const _currentDep = Symbol('currentDep')
 const _updateAll = Symbol('updateAll')
 const _mutateTree = Symbol('mutateTree')
@@ -75,7 +75,6 @@ const _follow = Symbol('follow')
 const _globalStyle = Symbol('globalStyle')
 const _globalRootNode = Symbol('globalRootNode')
 const _isVulnerable = Symbol.for('isVulnerable')
-const _dedupeList = Symbol.for('dedupeList')
 
 // used by Reify mixin
 const _force = Symbol.for('force')
@@ -605,7 +604,6 @@ module.exports = cls => class IdealTreeBuilder extends cls {
       .filter(edge => !bundled.has(edge.name) &&
         !(edge.to && edge.to.inShrinkwrap) &&
         (!edge.valid || !edge.to || this[_updateNames].includes(edge.name) ||
-          this[_dedupeList].has(edge.to.path) ||
           this[_isVulnerable](edge.to) ||
           node.isRoot && this[_explicitRequests].has(edge.name)))
   }
@@ -698,8 +696,7 @@ module.exports = cls => class IdealTreeBuilder extends cls {
     if (edge.to &&
         !edge.error &&
         !this[_updateNames].includes(edge.name) &&
-        !this[_isVulnerable](edge.to) &&
-        !this[_dedupeList].has(edge.to.path))
+        !this[_isVulnerable](edge.to))
       return []
 
     // top nodes should still get peer deps from their parent or fsParent
