+++ lib/arborist/index.js
@@ -32,7 +32,7 @@ const {homedir} = require('os')
 
 const mixins = [
   require('../tracker.js'),
-  require('./trimmer.js'),
+  require('./deduper.js'),
   require('./audit.js'),
   require('./build-ideal-tree.js'),
   require('./load-actual.js'),
